Tic Tac Toe/requirements.txt
Pillow==12.0.0
pygame==2.6.1
